+++ mesh/src/main/java/ca/mcmaster/cas/se2aa4/a2/mesh/adt/properties/ThicknessProperty.java
@@ -8,7 +8,7 @@ public class ThicknessProperty extends Property{
     public static final String KEY = "thickness";
 
     private static final String VALUE_FORMAT = "%f";
-    public ThicknessProperty(double x) {
+    public ThicknessProperty(float x) {
         super(KEY, String.format(VALUE_FORMAT, Util.precision(x)));
     }
