+++ mesh/src/main/java/ca/mcmaster/cas/se2aa4/a2/mesh/adt/vertex/Vertex.java
@@ -194,13 +194,13 @@ public class Vertex implements Indexable,Thickenable, IProperties, Renderable, C
     }
 
     @Override
-    public void setThickness(double x) {
+    public void setThickness(float x) {
         Property property = new ThicknessProperty(x);
         this.addProperty(property);
     }
 
     @Override
-    public double getThickness() {
+    public float getThickness() {
         return Util.extractThickness(this);
     }
 }
